requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
attrs==21.4.0
av==9.2.0
brotlipy==0.7.0
cachetools==5.0.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
compressai==1.2.0
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
docker-pycreds==0.4.0
flatbuffers==2.0
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
fvcore==0.1.5.post20220414
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.6
google-auth-oauthlib==0.4.6
grpcio==1.44.0
h5py==3.6.0
hydra-core==1.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio==2.19.0
importlib-metadata==4.11.3
install==1.3.5
iopath==0.1.9
jax==0.3.8
jaxlib==0.3.7
kiwisolver==1.4.2
lpips==0.1.4
Markdown==3.3.6
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
multidict==6.0.2
networkx==2.8
neuralcompression==0.2.1
ninja==1.10.2.3
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1649764630438/work
oauthlib==3.2.0
omegaconf==2.1.2
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
parameterized==0.8.1
pathtools==0.1.2
Pillow==9.0.1
portalocker==2.4.0
promise==2.3
protobuf==3.20.1
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyDeprecate==0.3.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.8
pypng==0.0.21
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
pytorch-lightning==1.6.2
pytorch-msssim==0.2.1
pytorchvideo==0.1.5
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scipy==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5
termcolor==1.1.0
torch==1.11.0
torchac==0.9.3
torchaudio==0.11.0
torchmetrics==0.8.1
torchvision==0.12.0
tqdm==4.64.0
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1650639997961/work
wandb==0.12.15
Werkzeug==2.1.2
yacs==0.1.8
yarl==1.7.2
zipp==3.8.0
